% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equil_points.R
\name{get_equi_points}
\alias{get_equi_points}
\title{Title}
\usage{
get_equi_points(results_file = OD_results, eq_threshold = 0.5, eq_var = coef)
}
\arguments{
\item{eq_var}{}
}
\value{

}
\description{
Title
}
