% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equil_points.R
\name{get_equil_points}
\alias{get_equil_points}
\title{get_equil_points}
\usage{
get_equil_points(results_file = OD_results, eq_threshold = 0.5, eq_var = coef)
}
\arguments{
\item{results_file}{: name of file where raw results have been stored. This file must be uploaded to the environment}

\item{eq_var}{: the method for calculating variance through time.   choose from coefficient of variance (coef), percent variance (perVar), or the functions in "cor" function from the stats package.}

\item{eq_threshold:}{the percent threshold for coefficient of variance to determine the point of stable equilibrium}
}
\value{

}
\description{
get_equil_points
}
\examples{
get_equil_points(results_file = ODE_results, eq_threshold = 0.01, eq_var = "perVar")
}
