% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{get_stats}
\usage{
get_stats(
  results_file = ODE_eq,
  test_parameters = c("sigmaBA" == "sigmaAB"),
  set_baseline = c(K = 200, lambda = 1, mu = 0.5)
)
}
\arguments{
\item{results_file}{: name of file where raw results have been stored. This file must be uploaded to the environment}

\item{test_parameters}{: vector of parameters tested.  if NA, all parameters will be tested apart from designated baseline.
additionally, parameters where values are to be matched during analysis can be specified with "etc == etc". parameters that are able to do defined in this way are: nu, sigma, and beta.}

\item{set_baseline}{: vector of parameters to be kept at baseline, if NA, all parameters combinations will be considered.}

\item{eq_threshold}{: the percent threshold for coefficient of variance to determine the point of stable equilibrium}
}
\value{
A list containing the following: details of the model including the date of
simulation and parameter combinations, raw results at equilibrium, correlation tests at equilibrium for each test parameter, and if ld is selected linkage disequilibrium at each parameter combination
}
\description{
get_stats
}
\examples{
get_stats(results_file = ODE_results,test_parameters = c("sigmaBA = sigmaAB"),set_baseline = c(K = 200, lambda = 1, mu = 0.5))
}
