% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot model}
\usage{
plot_model(
  results_file = ODE_results,
  colors = c("royalblue3", "orange", "red"),
  plot_file = "ODE_Plot.pdf"
)
}
\arguments{
\item{results_file}{: results}

\item{colors}{: a vector  of colors specifying the colour ramp}

\item{plot_file}{: file name to save plots}
}
\value{
PDF of plots of results over timesteps
}
\description{
Plot model
}
\examples{
plot_model(results_file = ODE_results, plot_file = "ODE_Plot.pdf", colours = c("royalblue3","orange", "red"))
}
