% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(
  endo_species = 2,
  endo_number = 2,
  parameters = params,
  tmax = 1000,
  int_plot = TRUE,
  save_results = c(TRUE, outfile = "ODE_results.rda")
)
}
\arguments{
\item{endo_species}{: number of endosymbiont species within the system}

\item{endo_number}{: number of endosymbionts per species within the system}

\item{parameters}{: a data frame containing all parameter combinations to run the model on}

\item{tmax}{: The maximum number of time steps}

\item{int_plot}{: if TRUE launches a shiny app for quick glances of results}

\item{save_results}{: if TRUE saves the results as a list with the selected file name.}
}
\value{
Shiny app of all results and /or  a list containing the details of the model,
a data frame of all parameter combinations,
and a list of model results for each parameter combination.
}
\description{
run_model
}
