% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{get_stats}
\usage{
get_stats(
  results_file = ODE_results,
  test_parameters = c("sigmaBA = sigmaAB"),
  set_baseline = c(K = 200, lambda = 1, mu = 0.5),
  cor_param_method = c("ld", "pearson"),
  outfile = "Stats.rda"
)
}
\arguments{
\item{results_file}{: name of file where raw results have been stored. This file must be uploaded to the environment}

\item{test_parameters}{: vector of parameters tested.  if NA, all parameters will be tested apart from designated baseline.
additionally, parameters where values are to be matched during analysis can be specified with "etc == etc". Parameters that are able to do defined in this way are: nu, sigma, and beta.}

\item{set_baseline}{: vector of parameters to be kept at baseline, if NA, all parameters combinations will be considered.}

\item{cor_param_method}{: methods of correlation calculations, options include ld, or
any method from the "cor" function in the stats package.}

\item{outfile}{: name of the file results will be saved.}
}
\value{
A list containing the following: details of the model including the date of
simulation and parameter combinations, raw results at equilibrium, correlation tests at equilibrium for each test parameter, and if ld is selected linkage disequilibrium at each paramter combination
}
\description{
get_stats
}
