% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_plot.R
\name{int_plot}
\alias{int_plot}
\title{int_plot}
\usage{
int_plot(
  parameters = params,
  endo_species = 2,
  endo_number = 1,
  tmax = 2500,
  kmax = NA,
  Colors = c("blue", "orange", "red")
)
}
\arguments{
\item{parameters}{: parameter combinations that will be simulated}

\item{endo_species}{: number of endosymbiont species no}

\item{endo_number}{: number of individuals per species}

\item{tmax}{: max number of timesteps}

\item{Colors}{:  vector of colours to be included in the graph}

\item{host_dem}{: if TRUE, host demographics (speciation, and extinction) will be included in the simulation.}
}
\value{
A shiny application line graph of number of hosts with intracellular endosymbionts within the system at each timestep at each parameter combination.
}
\description{
int_plot
}
\examples{
params <- set_parameters(two_species = TRUE,K = 200,lambda = 1,mu = 0.5,betaA =0.001,betaB = 0.001,sigmaA = 0.1,sigmaB = 0.1,sigmaAB = 1,sigmaBA = seq(0, 1, 0.1),nuA = 0.01,nuB = 0.01)
int_plot(parameters = params, endo_species = 2, endo_number = 1)
}
