% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot model}
\usage{
plot_model(
  results_file = ODE_results,
  plot_file = "ODE_Plot.pdf",
  Colors = c("blue", "orange", "red")
)
}
\arguments{
\item{results_file}{: results}

\item{plot_file}{: file name to save plots}

\item{Colors}{:  vector of colours to be included in the graph}
}
\value{
PDF of plots of results over time steps
}
\description{
Plot model
}
\examples{
plot_model(results_file = ODE_results, plot_file = "ODE_Plot.pdf", Colors = c("blue","orange", "red"))
}
