% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Create  area plots for simulations generated in "bugz"}
\usage{
plot_model(
  data = NULL,
  x_labs = expression(paste(sigma[A])),
  y_labs = "\% infected",
  legend_labs = "Infection\\n status \\n       AB",
  ind_var = "betaA",
  def_param_plot = NULL,
  colouring = c("dodgerblue4", "darkcyan", "cadetblue1"),
  alpha = 1,
  uninfected_colour = "white",
  def_line_colour = "grey",
  line_type = 1,
  legend_pos = "NULL",
  defaults = c(sigmaA == 0 & nuA == 0.005),
  xbreaks = NULL,
  ybreaks = NULL,
  titles = NULL,
  second_def_line = NULL,
  second_line_colour = NULL,
  tags = NULL,
  tag_pos = NULL,
  ticks = NULL
)
}
\arguments{
\item{data}{dataset}

\item{y_labs}{Label for y axis}

\item{ind_var}{desired independent variable}

\item{def_param_plot}{default parameters to plot}

\item{colouring}{vector for the colour gradient desired in the plot}

\item{def_line_colour}{the colour of the line representing the default parameter}

\item{line_type}{line type of default parameter}

\item{legend_pos}{legend position}

\item{defaults}{the default parameter values for the dataset}

\item{xbreaks}{the breaks for the tick marks and number on y axis}

\item{ybreaks}{the breaks for the tick marks and numbers on the x axis}

\item{titles}{title for the graph}

\item{second_def_line}{to add another line displaying a second default parameter to the plot}

\item{second_line_colour}{colour for this second default parameter line}

\item{tags}{tags for the plot}

\item{Legend_labs}{Label for legend}
}
\value{
a 100% stacked area plot of results
}
\description{
Create  area plots for simulations generated in "bugz"
}
